scatter qt running from terminal
